+++ filament/src/details/Stream.cpp
@@ -33,7 +33,6 @@ using namespace backend;
 
 struct Stream::BuilderDetails {
     void* mStream = nullptr;
-    intptr_t mExternalTextureId = 0;
     uint32_t mWidth = 0;
     uint32_t mHeight = 0;
 };
@@ -52,11 +51,6 @@ Stream::Builder& Stream::Builder::stream(void* stream) noexcept {
     return *this;
 }
 
-Stream::Builder& Stream::Builder::stream(intptr_t externalTextureId) noexcept {
-    mImpl->mExternalTextureId = externalTextureId;
-    return *this;
-}
-
 Stream::Builder& Stream::Builder::width(uint32_t width) noexcept {
     mImpl->mWidth = width;
     return *this;
@@ -68,11 +62,6 @@ Stream::Builder& Stream::Builder::height(uint32_t height) noexcept {
 }
 
 Stream* Stream::Builder::build(Engine& engine) {
-    if (!ASSERT_PRECONDITION_NON_FATAL(!mImpl->mStream || !mImpl->mExternalTextureId,
-            "One and only one of the stream or external texture can be specified")) {
-        return nullptr;
-    }
-
     return upcast(engine).createStream(*this);
 }
 
@@ -80,21 +69,14 @@ Stream* Stream::Builder::build(Engine& engine) {
 
 FStream::FStream(FEngine& engine, const Builder& builder) noexcept
         : mEngine(engine),
-          mStreamType(
-            builder->mExternalTextureId ? StreamType::TEXTURE_ID :
-            (builder->mStream ? StreamType::NATIVE : StreamType::ACQUIRED)
-          ),
+          mStreamType(builder->mStream ? StreamType::NATIVE : StreamType::ACQUIRED),
           mNativeStream(builder->mStream),
-          mExternalTextureId(builder->mExternalTextureId),
           mWidth(builder->mWidth),
           mHeight(builder->mHeight) {
 
     if (mNativeStream) {
         // Note: this is a synchronous call. On Android, this calls back into Java.
         mStreamHandle = engine.getDriverApi().createStreamNative(mNativeStream);
-    } else if (mExternalTextureId) {
-        mStreamHandle = engine.getDriverApi().createStreamFromTextureId(
-                mExternalTextureId, mWidth, mHeight);
     } else {
         mStreamHandle = engine.getDriverApi().createStreamAcquired();
     }
@@ -122,28 +104,6 @@ void FStream::setDimensions(uint32_t width, uint32_t height) noexcept {
         FFence::waitAndDestroy(mEngine.createFence(FFence::Type::SOFT), Fence::Mode::FLUSH);
     }
     mEngine.getDriverApi().setStreamDimensions(mStreamHandle, mWidth, mHeight);
-}
-
-void FStream::readPixels(uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
-        backend::PixelBufferDescriptor&& buffer) noexcept {
-    if (getStreamType() == StreamType::TEXTURE_ID) {
-        // this works only on external texture id streams
-
-        const size_t sizeNeeded = PixelBufferDescriptor::computeDataSize(
-                buffer.format, buffer.type,
-                buffer.stride ? buffer.stride : width,
-                buffer.top + height,
-                buffer.alignment);
-
-        if (!ASSERT_POSTCONDITION_NON_FATAL(buffer.size >= sizeNeeded,
-                "buffer.size too small %u bytes, needed %u bytes", buffer.size, sizeNeeded)) {
-            return;
-        }
-
-        FEngine::DriverApi& driver = mEngine.getDriverApi();
-        driver.readStreamPixels(mStreamHandle,
-                xoffset, yoffset, width, height, std::move(buffer));
-    }
 }
 
 int64_t FStream::getTimestamp() const noexcept {